WIP: move to more callback like structure
